Fix missing menu icon
This has to do with some change to react-navigation v3 and passing
navigation options when creating the navigator. I don't couldn't get
that working and worked around it using the static navigationOptions.
I'm not sure if this goes against the intent of passing it during create
because I don't understand the difference between the for navigators
yet.

See https://stackoverflow.com/a/56096215/34315 and https://github.com/react-navigation/react-navigation/issues/1122\#issuecomment-513244470